LSR-851 Url generation back to model
